User Settings stored in the Realm DB - success.

+++ app/src/main/java/com/wipro/wipro_music_player/model/FavouriteSongModel.java
@@ -5,7 +5,7 @@ import io.realm.annotations.PrimaryKey;
 
 public class FavouriteSongModel extends RealmObject {
     @PrimaryKey
-    private int id;
+    private long id;
     private String songTitle;
     private String songArtist;
     private String songPath;
@@ -14,14 +14,14 @@ public class FavouriteSongModel extends RealmObject {
 
     public FavouriteSongModel() {}
 
-    public int getId() { return id; }
+    public long getId() { return id; }
     public String getSongTitle() { return songTitle; }
     public String getSongArtist() { return songArtist; }
     public String getSongPath() { return songPath; }
     public long getSongLength() { return songLength; }
     public double getSongSize() { return songSize; }
 
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
     public void setSongTitle(String songTitle) { this.songTitle = songTitle; }
     public void setSongArtist(String songArtist) { this.songArtist = songArtist; }
     public void setSongPath(String songPath) { this.songPath = songPath; }
